Changed the way imageSwipe Component is organized
